reorganize location picker widgets for readability
